requirements.txt
asgiref==3.2.10
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
cryptography==3.2.1
defusedxml==0.6.0
Django==3.1.1
django-allauth==0.44.0
django-tinymce4-lite==1.8.0
easy-thumbnails==2.7.1
idna==2.10
jsmin==2.2.2
oauthlib==3.1.0
Pillow==8.0.1
psycopg2==2.8.6
pycparser==2.20
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2020.4
redis==3.5.3
requests==2.25.0
requests-oauthlib==1.3.0
six==1.15.0
sorl-thumbnail==12.7.0
sqlparse==0.4.1
urllib3==1.26.2
Wand==0.6.5
